move all subgraph routines into the same file
