+++ python/dgl/core.py
@@ -1,6 +1,6 @@
 """Implementation for core graph computation."""
 
-from .base import DGLError
+from .base import DGLError, is_all
 from . import function as fn
 from .udf import NodeBatch, EdgeBatch
 from . import ops
@@ -9,17 +9,39 @@ def is_builtin(func):
     return isinstance(func, fn.BuiltinFunction)
 
 def invoke_node_udf(graph, nodes, ntype, func, ndata):
-    nbatch = NodeBatch(graph, nodes, ndata, ntype)
+    if is_all(nodes):
+        nodes = graph.nodes()
+    nbatch = NodeBatch(graph, nodes, ntype, ndata)
     return func(nbatch)
 
 def invoke_udf_reduce(graph, func, edata=None):
     pass
 
-def invoke_edge_udf(graph, func):
-    pass
+def invoke_edge_udf(graph, eid, etype, func):
+    if is_all(eid):
+        u, v, eid = graph.edges(form='all', etype=etype)
+    else:
+        u, v = graph.find_edges(eid, etype=etype)
+    etid = graph.get_etype_id(etype)
+    stid, dtid = graph._graph.metagraph.find_edge(etid)
+    srcdata = graph._node_frames[stid].subframe(u)
+    dstdata = graph._node_frames[dtid].subframe(v)
+    edata = graph._edge_frames[etid].subframe(eid)
+    ebatch = EdgeBatch(graph, eid, etype, srcdata, edata, dstdata)
+    return func(ebatch)
 
 def invoke_gsddmm(graph, func):
-    pass
+    alldata = [graph.srcdata, graph.dstdata, graph.edata]
+    if isinstance(func, fn.BinaryMessageFunction):
+        x = alldata[func.lhs][func.lhs_field]
+        y = alldata[func.rhs][func.rhs_field]
+        op = getattr(ops, func.name)
+        z = op(graph, x, y)
+    else:
+        x = alldata[func.target][func.in_field]
+        op = getattr(ops, func.name)
+        z = op(graph, x)
+    return {func.out_field : z}
 
 def invoke_gspmm(graph, mfunc, rfunc, *, srcdata=None, dstdata=None, edata=None):
     # sanity check
@@ -46,7 +68,7 @@ def invoke_gspmm(graph, mfunc, rfunc, *, srcdata=None, dstdata=None, edata=None)
         z = op(graph, x)
     return {rfunc.out_field : z}
 
-def message_passing(g, mfunc, rfunc, afunc):
+def message_passing(g, mfunc, rfunc):
     if g.number_of_edges() == 0:
         # No message passing is triggered.
         return
@@ -65,8 +87,4 @@ def message_passing(g, mfunc, rfunc, afunc):
         else:
             edata.update(g.edata)  # incorporate original edge features
             ndata = invoke_udf_reduce(g, rfunc, edata=edata)
-    # apply phase
-    if afunc is not None:
-        ndata.update(g.dstdata)  # incorporate original node features
-        ndata = invoke_node_udf(g, ALL, afunc, ndata)
     return ndata
