+++ src/graph/heterograph_capi.cc
@@ -648,7 +648,7 @@ DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroGetFormatGraph")
     *rv = HeteroGraphRef(hgptr);
 });
 
-DGL_REGISTER_GLOBAL("transform._CAPI_DGLInSubgraph")
+DGL_REGISTER_GLOBAL("subgraph._CAPI_DGLInSubgraph")
 .set_body([] (DGLArgs args, DGLRetValue *rv) {
     HeteroGraphRef hg = args[0];
     const auto& nodes = ListValueToVector<IdArray>(args[1]);
@@ -657,7 +657,7 @@ DGL_REGISTER_GLOBAL("transform._CAPI_DGLInSubgraph")
     *rv = HeteroGraphRef(ret);
   });
 
-DGL_REGISTER_GLOBAL("transform._CAPI_DGLOutSubgraph")
+DGL_REGISTER_GLOBAL("subgraph._CAPI_DGLOutSubgraph")
 .set_body([] (DGLArgs args, DGLRetValue *rv) {
     HeteroGraphRef hg = args[0];
     const auto& nodes = ListValueToVector<IdArray>(args[1]);
